feat(menubar): when switching between fragments, make the menu bar remove items or change text
