+++ app/src/main/java/com/chesire/malime/view/MainActivity.kt
@@ -5,11 +5,9 @@ import android.net.Uri
 import android.os.Bundle
 import android.support.customtabs.CustomTabsIntent
 import android.support.design.widget.BottomNavigationView
-import android.support.design.widget.Snackbar
 import android.support.v4.app.Fragment
 import android.support.v7.app.AlertDialog
 import android.support.v7.app.AppCompatActivity
-import android.view.Menu
 import android.view.MenuItem
 import com.chesire.malime.R
 import com.chesire.malime.util.PeriodicUpdateHelper
@@ -19,7 +17,6 @@ import com.chesire.malime.view.login.LoginActivity
 import com.chesire.malime.view.manga.MangaFragment
 import com.chesire.malime.view.preferences.PrefActivity
 import com.chesire.malime.view.search.SearchFragment
-import timber.log.Timber
 
 class MainActivity : AppCompatActivity() {
     private val sharedPref: SharedPref by lazy {
@@ -81,11 +78,6 @@ class MainActivity : AppCompatActivity() {
         super.onSaveInstanceState(outState)
     }
 
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.menu_options, menu)
-        return true
-    }
-
     override fun onOptionsItemSelected(item: MenuItem?): Boolean {
         when {
             item?.itemId == R.id.menu_options_view_profile -> {
@@ -116,39 +108,12 @@ class MainActivity : AppCompatActivity() {
 
                 return true
             }
-            item?.itemId == R.id.menu_options_filter -> {
-                spawnFilterDialog()
-                return true
-            }
             item?.itemId == R.id.menu_options_sort -> {
                 spawnSortDialog()
                 return true
             }
             else -> return super.onOptionsItemSelected(item)
         }
-    }
-
-    private fun spawnFilterDialog() {
-        val states = sharedPref.getAnimeFilter()
-        AlertDialog.Builder(this)
-            .setTitle(R.string.filter_dialog_title)
-            .setMultiChoiceItems(R.array.anime_states, states, { _, which, isChecked ->
-                states[which] = isChecked
-            })
-            .setPositiveButton(android.R.string.ok, { _, _ ->
-                if (states.all { !it }) {
-                    Timber.w("User tried to set all filter states to false")
-                    Snackbar.make(
-                        findViewById(R.id.activity_main_layout),
-                        R.string.filter_must_select,
-                        Snackbar.LENGTH_LONG
-                    ).show()
-                } else {
-                    sharedPref.setAnimeFilter(states)
-                }
-            })
-            .setNegativeButton(android.R.string.cancel, null)
-            .show()
     }
 
     private fun spawnSortDialog() {
