+++ app/src/main/res/layout/fragment_maldisplay.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/maldisplay_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
