+++ app/src/main/res/menu/menu_search.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/menu_options_view_profile"
+        android:title="@string/options_view_profile" />
+    <item
+        android:id="@+id/menu_options_settings"
+        android:title="@string/options_settings" />
+    <item
+        android:id="@+id/menu_options_log_out"
+        android:title="@string/options_log_out" />
+</menu>
